// frontend/src/api/models/CreateResidentDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateResidentDto = {
    apartmentId?: string;
    fullName: string;
    phone: string;
    password: string;
    email: string;
    role?: string;
    temporaryStatus?: boolean;
    idNumber: string;
    birthDate: string;
};